refactor: consolidate Android and JVM calling code in logic module
