+++ logic/src/darwinMain/kotlin/com/wire/kalium/logic/feature/call/MediaManagerServiceImpl.kt
@@ -18,9 +18,12 @@
 
 package com.wire.kalium.logic.feature.call
 
+import com.wire.kalium.logic.util.PlatformContext
 import kotlinx.coroutines.flow.StateFlow
 
-actual class MediaManagerServiceImpl : MediaManagerService {
+actual class MediaManagerServiceImpl(
+    appContext: PlatformContext
+) : MediaManagerService {
     override fun turnLoudSpeakerOn() {
         TODO("Not yet implemented")
     }
