+++ logic/src/darwinMain/kotlin/com/wire/kalium/logic/util/PlatformContext.kt
@@ -16,17 +16,6 @@
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
 
-package com.wire.kalium.logic.feature.call.scenario
+package com.wire.kalium.logic.util
 
-import com.sun.jna.Pointer
-import com.wire.kalium.calling.callbacks.VideoReceiveStateHandler
-import com.wire.kalium.logic.callingLogger
-
-class OnParticipantsVideoStateChanged : VideoReceiveStateHandler {
-    override fun onVideoReceiveStateChanged(conversationId: String, userId: String, clientId: String, state: Int, arg: Pointer?) {
-        callingLogger.i(
-            "[onVideoReceiveStateChanged] - conversationId: $conversationId | userId: $userId clientId: $clientId" +
-                    " | state: $state"
-        )
-    }
-}
+actual class PlatformContext
