+++ logic/src/jvmMain/kotlin/com/wire/kalium/logic/feature/call/FlowManagerServiceImpl.kt
@@ -20,9 +20,12 @@ package com.wire.kalium.logic.feature.call
 
 import com.wire.kalium.logic.data.id.ConversationId
 import com.wire.kalium.logic.kaliumLogger
+import com.wire.kalium.logic.util.PlatformContext
 import com.wire.kalium.logic.util.PlatformView
 
-actual class FlowManagerServiceImpl : FlowManagerService {
+actual class FlowManagerServiceImpl(
+    appContext: PlatformContext
+) : FlowManagerService {
     override suspend fun setVideoPreview(conversationId: ConversationId, view: PlatformView) {
         kaliumLogger.w("setVideoPreview for JVM but not supported yet.")
     }
