--- logic/src/jvmMain/kotlin/com/wire/kalium/logic/feature/call/scenario/OnActiveSpeakers.kt
@@ -1,44 +1,0 @@
-/*
- * Wire
- * Copyright (C) 2023 Wire Swiss GmbH
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see http://www.gnu.org/licenses/.
- */
-
-package com.wire.kalium.logic.feature.call.scenario
-
-import com.sun.jna.Pointer
-import com.wire.kalium.calling.callbacks.ActiveSpeakersHandler
-import com.wire.kalium.calling.types.Handle
-import com.wire.kalium.logic.data.call.CallActiveSpeakers
-import com.wire.kalium.logic.data.call.CallRepository
-import com.wire.kalium.logic.data.id.QualifiedIdMapper
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.json.Json
-
-class OnActiveSpeakers(
-    private val callRepository: CallRepository,
-    private val qualifiedIdMapper: QualifiedIdMapper
-) : ActiveSpeakersHandler {
-
-    override fun onActiveSpeakersChanged(inst: Handle, conversationId: String, data: String, arg: Pointer?) {
-        val callActiveSpeakers = Json.decodeFromString<CallActiveSpeakers>(data)
-        val conversationIdWithDomain = qualifiedIdMapper.fromStringToQualifiedID(conversationId)
-
-        callRepository.updateParticipantsActiveSpeaker(
-            conversationId = conversationIdWithDomain.toString(),
-            activeSpeakers = callActiveSpeakers
-        )
-    }
-}
