--- logic/src/jvmMain/kotlin/com/wire/kalium/logic/feature/call/scenario/OnNetworkQualityChanged.kt
@@ -1,41 +1,0 @@
-/*
- * Wire
- * Copyright (C) 2023 Wire Swiss GmbH
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see http://www.gnu.org/licenses/.
- */
-
-package com.wire.kalium.logic.feature.call.scenario
-
-import com.sun.jna.Pointer
-import com.wire.kalium.calling.callbacks.NetworkQualityChangedHandler
-import com.wire.kalium.logger.obfuscateId
-import com.wire.kalium.logic.callingLogger
-
-class OnNetworkQualityChanged : NetworkQualityChangedHandler {
-
-    override fun onNetworkQualityChanged(
-        conversationId: String,
-        userId: String?,
-        clientId: String?,
-        quality: Int,
-        roundTripTimeInMilliseconds: Int,
-        upstreamPacketLossPercentage: Int,
-        downstreamPacketLossPercentage: Int,
-        arg: Pointer?
-    ) {
-        // Not yet implemented
-        callingLogger.i("[OnNetworkQualityChanged] - ConversationId: ${conversationId.obfuscateId()} | UserId: $userId | Quality: $quality")
-    }
-}
